revert hacks

+++ js/api_tests/src/deadbeef/player_controller.js
@@ -14,6 +14,7 @@ const close = promisify(fs.close);
 
 const mkdirp = promisify(require('mkdirp'));
 const rimraf = promisify(require('rimraf'));
+const tmpdir = promisify(require('tmp').dir);
 
 class PlayerController
 {
@@ -33,7 +34,6 @@ class PlayerController
         if (!this.paths.profileDir)
             await this.initProfile();
 
-        await this.removeProfileDir();
         await this.installPlugins();
         await this.writePlayerSettings();
         await this.writePluginSettings(pluginSettings);
@@ -43,19 +43,15 @@ class PlayerController
     async stop()
     {
         this.stopProcess();
+        await this.removeTempFiles();
     }
 
     async getLog()
     {
-        let data = '';
-
-        if (this.paths.configFile)
-            data = data + '\nconfig file:\n' + await readFile(this.paths.configFile, 'utf8');
-
         if (this.paths.logFile)
-            data = data + '\nrun log:\n' + await readFile(this.paths.logFile, 'utf8');
+            return await readFile(this.paths.logFile, 'utf8');
 
-        return data;
+        return null;
     }
 
     async findPlayerBinary()
@@ -86,7 +82,7 @@ class PlayerController
 
     async initProfile()
     {
-        const profileDir = path.join(this.config.testsRootDir, 'tmp');
+        const profileDir = await tmpdir({ prefix: 'beefweb-api-tests-' });
         const configDir = path.join(profileDir, '.config/deadbeef');
         const libDir = path.join(profileDir, '.local/lib/deadbeef');
         const configFile = path.join(configDir, 'config');
@@ -135,9 +131,10 @@ class PlayerController
         }
     }
 
-    async removeProfileDir()
+    async removeTempFiles()
     {
-        await rimraf(this.paths.profileDir);
+        if (this.paths.profileDir)
+            await rimraf(this.paths.profileDir);
     }
 
     async startProcess(environment)
